Use Virtus 1.0

+++ lib/barometer/data/attributes/integer.rb
@@ -1,9 +1,9 @@
 module Barometer
   module Data
     module Attribute
-      class Float < Virtus::Attribute
+      class Integer < Virtus::Attribute
         def coerce(value, *args)
-          value ? value.to_f : default_value.call
+          value ? value.to_i : default_value.call
         end
       end
     end
